Move `get_available_cpu_count()` to config.py
The fact that operator_utils.py imported numpy, scipy, and other
libraries made it impossible to import `get_available_cpu_count()`
alone. This in turn prevented using it to set things before NumPy and
other libraries were loaded.
