+++ src/openfermion/linalg/linear_qubit_operator.py
@@ -20,7 +20,8 @@ import numpy.linalg
 import scipy.sparse
 import scipy.sparse.linalg
 
-from openfermion.utils.operator_utils import count_qubits, get_available_cpu_count
+from openfermion.utils.operator_utils import count_qubits
+from openfermion.config import get_available_cpu_count
 
 
 class LinearQubitOperatorOptions:
